Thu, Jan 11, 2024, 12:15 PM +02:00

+++ tip/tip.py
@@ -1,7 +1,7 @@
 def main():
     dollars = dollars_to_float(input("How much was the meal? "))
     percent = percent_to_float(input("What percentage would you like to tip? "))
-    tip = int(dollars) * int(percent)
+    tip = int(dollars) * float(percent)
     print(f"Leave ${tip:.2f}")
